add delete 弹窗

+++ src/views/table/complex-table.vue
@@ -205,13 +205,24 @@ export default {
       return a < b
     },
     handleDelete(id) {
-      deleteInfo({ 'id': id }).then(response => {
-        this.$message({
-          showClose: true,
-          message: '删除成功',
-          type: 'success'
+      this.$confirm('此操作将永久删除, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deleteInfo({'id': id}).then(response => {
+          this.$message({
+            showClose: true,
+            message: '删除成功',
+            type: 'success'
+          })
+          this.getList()
         })
-        this.getList()
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        })
       })
     },
     edit_visiable(row) {
